refacto: rename nlptest -> nhelper
